Refactor: change name of domain objects
