+++ src/main/java/jupgo/jupgoserver/domain/User.java
@@ -1,4 +1,4 @@
-package jupgo.jupgoserver.dto;
+package jupgo.jupgoserver.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -6,7 +6,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Entity
-public class UserDto {
+public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
